Resolves #78, #63. Implement `cycle` by fixing env binding error.
This was a really nasty one so it definitely deserves a description.

The
```
eval env (MappyClosure args body env') = Right $ MappyClosure args body (env ++ env')
```
line that was removed from Executor.hs was wrong. I've always been a
little suspicious of it, but it's clear what the issue was now.

Closures should only care about the environment they were initially
bound in, the above line caused this to not be true (every time they
were executed they got a new environment -- the current one -- added to them).

`extendEnvironment` was also at fault. Previously the _closure's_
environment was being used to evaluate arguments that were being passed
into function evaluation. This worked, coincidentally, with the above
line. This, however, was also wrong.

The bottom line is
> The arguments that are passed to a function must first be evaluated in
> the CURRENT environment, not the closed one.